+++ packages/sdk/src/common/sanitize-uri.ts
@@ -1,7 +1,17 @@
 
 export function sanitizeUri(contractPrefix: string, uriRaw: string): string {
-	if (!uriRaw.startsWith(contractPrefix)) {
-		throw new Error(`uri must starts with: ${contractPrefix}`)
+	const fixedContractPrefix = fixContractPrefix(contractPrefix)
+	if (!uriRaw.startsWith(fixedContractPrefix)) {
+		throw new Error(`uri must start with: ${fixedContractPrefix}`)
 	}
-	return uriRaw.slice(contractPrefix.length) || ""
+	return uriRaw.slice(fixedContractPrefix.length) || ""
+}
+
+/**
+ * Workaround for older contracts where contract uri is set to http link
+ */
+function fixContractPrefix(contractPrefix: string): string {
+	return contractPrefix
+		.replace("https://ipfs.daonomic.com", "ipfs:/")
+		.replace("https://ipfs.rarible.com", "ipfs:/")
 }