+++ packages/sdk/src/order/asset-type-to-struct.test.ts
@@ -1,5 +1,5 @@
 import { toAddress, toBigNumber, toBinary } from "@rarible/types"
-import { createGanacheProvider } from "@rarible/ethereum-sdk-test-common"
+import { createGanacheProvider } from "@rarible/ethereum-sdk-test-common/build/create-ganache-provider"
 import { createTestProviders } from "../common/create-test-providers"
 import { assetTypeToStruct } from "./asset-type-to-struct"