+++ packages/sdk/src/order/fill-order/crypto-punks.test.ts
@@ -1,7 +1,8 @@
 import { randomWord, toAddress, toBigNumber, toBinary, ZERO_ADDRESS } from "@rarible/types"
 import { Web3Ethereum } from "@rarible/web3-ethereum"
 import Web3 from "web3"
-import { awaitAll, createGanacheProvider } from "@rarible/ethereum-sdk-test-common"
+import { awaitAll } from "@rarible/ethereum-sdk-test-common"
+import { createGanacheProvider} from "@rarible/ethereum-sdk-test-common/build/create-ganache-provider"
 import { sentTx, simpleSend } from "../../common/send-transaction"
 import { getEthereumConfig } from "../../config"
 import { deployTestErc20 } from "../contracts/test/test-erc20"
@@ -90,7 +91,7 @@ describe("fillOrder", () => {
 		await sentTx(it.punksMarket.methods.allInitialOwnersAssigned(), {from: sender1Address})
 
 	})
-	test("should fill order with crypto punks asset", async () => {
+	test("should fill order (buy) with crypto punks asset", async () => {
 		//Mint crypto punks
 		const punkId = 43
 		const punkPrice = 10
@@ -123,7 +124,7 @@ describe("fillOrder", () => {
 		}
 
 		const finalOrder = { ...left, signature: toBinary("0x") }
-		const tx = await filler.fill({ order: finalOrder, amount: 1 })
+		const tx = await filler.buy({ order: finalOrder, amount: 1 })
 		await tx.wait()
 
 		await retry(5, 500, async () => {
@@ -133,7 +134,7 @@ describe("fillOrder", () => {
 		})
 	})
 
-	test("should fill bid with crypto punks asset", async () => {
+	test("should accept bid with crypto punks asset", async () => {
 		const punkId = 50
 		const punkPrice = 10
 		//Mint punks
@@ -165,7 +166,7 @@ describe("fillOrder", () => {
 		}
 
 		const finalOrder = { ...left, signature: toBinary("0x") }
-		const tx = await filler.fill({ order: finalOrder, amount: 1 })
+		const tx = await filler.acceptBid({ order: finalOrder, amount: 1 })
 		await tx.wait()
 
 		await retry(5, 500, async () => {