+++ packages/test-common/src/index.ts
@@ -1,6 +1,5 @@
 export { createE2eProvider, createE2eWallet } from "./create-e2e-provider"
 export { testTypedSignature } from "./test-typed-signature"
 export { testPersonalSign } from "./test-personal-sign"
-export { createGanacheProvider } from "./create-ganache-provider"
 export { testSimpleContract } from "./test-contract"
 export { awaitAll } from "./await-all"