+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.19](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.18...v0.12.19) (2021-12-10)
+
+
+### Bug Fixes
+
+* update CryptoPunks transfer proxy on Rinkeby. ([f7cc59f](https://github.com/rarible/protocol-ethereum-sdk/commit/f7cc59f42f765feb8aa389bd3dd76430971bb53a))
+
+
+
+
+
+## [0.12.18](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.17...v0.12.18) (2021-12-10)
+
+
+### Features
+
+* replace fill order with buy & acceptBid orders ([062a142](https://github.com/rarible/protocol-ethereum-sdk/commit/062a1425339889efa3ab933b333c47f39275a344))
+
+
+
+
+
+## [0.12.17](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.16...v0.12.17) (2021-12-09)
+
+
+### Bug Fixes
+
+* allow to update CryptoPunk orders with ETH. ([c25f855](https://github.com/rarible/protocol-ethereum-sdk/commit/c25f85515c5058392e5bfba0dc71a79c73c33f7a))
+
+
+
+
+
+## [0.12.16](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.15...v0.12.16) (2021-12-08)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.12.15](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.14...v0.12.15) (2021-12-07)
+
+
+### Bug Fixes
+
+* eth asset type to bid request was added ([89afbfc](https://github.com/rarible/protocol-ethereum-sdk/commit/89afbfc6a07226249644586dbefbe4610258b537))
+
+
+
+
+
+## [0.12.14](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.12...v0.12.14) (2021-12-07)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.12.12](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.11...v0.12.12) (2021-12-07)
+
+**Note:** Version bump only for package @rarible/protocol-ethereum-sdk
+
+
+
+
+
+## [0.12.11](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.10...v0.12.11) (2021-12-06)
+
+
+### Bug Fixes
+
+* use better dependencies for web3 ([4b0493f](https://github.com/rarible/protocol-ethereum-sdk/commit/4b0493f4684c95bb39ec4063b2bd588f03e1a0f7))
+* workaround for old contracts base uri ([ef06a7d](https://github.com/rarible/protocol-ethereum-sdk/commit/ef06a7db3af2135a1676bc4b2de1b6d176f314e2))
+
+
+
+
+
 ## [0.12.10](https://github.com/rarible/protocol-ethereum-sdk/compare/v0.12.9...v0.12.10) (2021-12-03)